added the ability for the script to create a directory to avoid an OSError
